requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
# Для создания графиков
matplotlib>=3.7.0
numpy>=1.24.0

# Дополнительные зависимости (если нужно)
# python-dotenv>=1.0.0  # Для .env файлов